fastapi/requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
beautifulsoup4
blinker
build
CacheControl
cachetools
certifi
charset-normalizer
cleo
click
cloudevents
colorama
contourpy
crashtest
cycler
dataclasses-json
deprecation
distlib
dulwich
entrypoints
faiss-cpu
Faker
fastapi
favicon
filelock
Flask
fonttools
frozenlist
fsspec
functions-framework
gitdb
GitPython
greenlet
h11
htbuilder
huggingface-hub
idna
importlib-metadata
installer
itsdangerous
jaraco.classes
Jinja2
joblib
jsonschema
keyring
kiwisolver
langchain
langsmith
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
marshmallow
matplotlib
mdurl
more-itertools
mpmath
msgpack
multidict
mypy-extensions
numexpr
numpy
openai
openapi-schema-pydantic
packaging
pandas
pdf2image
pexpect
Pillow
pkginfo
platformdirs
poetry
poetry-core
poetry-plugin-export
protobuf
ptyprocess
pyarrow
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
pymdown-extensions
Pympler
PyMuPDF
PyMuPDFb
pyparsing
PyPDF2
pyproject_hooks
pyrsistent
pytesseract
python-dateutil
python-dotenv
pytz
pywin32-ctypes
PyYAML
rapidfuzz
regex
requests
requests-toolbelt
rich
safetensors
semver
sentencepiece
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
st-annotated-text
starlette
streamlit
streamlit-camera-input-live
streamlit-card
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-toggle-switch
streamlit-vertical-slider
sympy
tabulate
tenacity
termcolor
textbase-client
threadpoolctl
tiktoken
tokenizers
toml
tomlkit
toolz
tornado
tqdm
transformers
trove-classifiers
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
validators
virtualenv
Wand
watchdog
Werkzeug
yarl
yaspin
zipp
scikit-learn
uvicorn[standard]
